requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comtypes==1.4.11
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-swagger==0.2.25
filelock==3.18.0
fsspec==2025.7.0
gradio_client==1.11.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
jwcrypto==1.5.6
kombu==5.5.4
lxml==6.0.1
packaging==25.0
passlib==1.7.4
pillow==11.3.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-jwt==4.1.0
python-multipart==0.0.20
python-pptx==1.0.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.3
rsa==4.9.1
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
xlsxwriter==3.2.5
pdf2image==1.17.0
PyMuPDF==1.24.14
